+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -334,7 +334,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWB8CGCAXEH1E8BF53DFHX6B"
+        "01KWBJ4807EJ1NQQPQ7Z6SNYKS"
     )
     r"""The unique identifier of the data source"""