+++ packages/orq-rc/src/orq_ai_sdk/models/evaluatorresponsehttp.py
@@ -199,9 +199,9 @@ class EvaluatorResponseHTTP(BaseModel):
 
     key: str
 
-    created: Optional[str] = "2026-06-30T03:15:04.225Z"
+    created: Optional[str] = "2026-06-30T06:05:21.250Z"
 
-    updated: Optional[str] = "2026-06-30T03:15:04.225Z"
+    updated: Optional[str] = "2026-06-30T06:05:21.250Z"
 
     updated_by_id: OptionalNullable[str] = UNSET