+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KWB8CGC7BH51GR6SGGT37BMR"
+        "01KWBJ4804E4XTXX4NPWH7JKZG"
     )
     r"""The unique identifier of the data source"""