+++ packages/orq-rc/src/orq_ai_sdk/models/retrievedatapointop.py
@@ -809,7 +809,7 @@ class RetrieveDatapointEvaluations3(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-30T03:15:25.273Z")
+    ] = parse_datetime("2026-06-30T06:05:40.314Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -954,7 +954,7 @@ class RetrieveDatapointEvaluations2(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-30T03:15:25.272Z")
+    ] = parse_datetime("2026-06-30T06:05:40.313Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1099,7 +1099,7 @@ class RetrieveDatapointEvaluations1(BaseModel):
         pydantic.Field(
             deprecated="warning: ** DEPRECATED ** - This will be removed in a future release, please migrate away from it as soon as possible."
         ),
-    ] = parse_datetime("2026-06-30T03:15:25.271Z")
+    ] = parse_datetime("2026-06-30T06:05:40.312Z")
     r"""Deprecated. The date and time the item was reviewed"""
 
     @model_serializer(mode="wrap")
@@ -1212,7 +1212,7 @@ class RetrieveDatapointResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-06-30T03:15:02.428Z")
+    updated: Optional[datetime] = parse_datetime("2026-06-30T06:05:20.131Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")