+++ packages/orq-rc/RELEASES.md
@@ -9079,3 +9079,13 @@ Based on:
 - [python v4.11.0-rc.57] packages/orq-rc
 ### Releases
 - [PyPI v4.11.0-rc.57] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.57 - packages/orq-rc
+
+## 2026-06-30 06:07:26
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.786.1 (2.913.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.11.0-rc.58] packages/orq-rc
+### Releases
+- [PyPI v4.11.0-rc.58] https://pypi.org/project/orq-ai-sdk/4.11.0-rc.58 - packages/orq-rc